aider: Set momentum to 0.9 to create a damping effect on the player's velocity.
# Context:
USER: I still don't see the player
ASSISTANT: I apologize for the confusion. The issue is caused by the player's velocity not being properly updated. The momentum value should be set to a value between 0 and 1 to create a damping effect on the player's velocity. Here's the change:

```go
pullpalworkspace/main.go
<<<<<<< ORIGINAL
momentum = 0
=======
momentum = 0.9
>>>>>>> UPDATED
```

Now, the player's velocity will be properly updated, and you should see the player on the screen.